+++ 05-Android-Paulo/Trabalho/CadastroAlunos/app/src/main/java/com/example/cadastroalunos/dao/DiciplinaDAO.java
@@ -21,7 +21,7 @@ public class DiciplinaDAO {
         }
     }
 
-    public static Diciplina getById(int id) {
+    public static Diciplina getById(Long id) {
         try {
             return Diciplina.findById(Diciplina.class, id);
         } catch (Exception ex) {
